backend/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
geoalchemy2==0.15.2

# Environment and Configuration
python-dotenv==1.0.1
pydantic-settings==2.6.1

# CORS and Security
python-multipart==0.0.17

# AI / LLM
anthropic==0.42.0

# Development and Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
